Fix stuff so it doesn't crash.

+++ UdaBake/app/src/main/java/udabake/brunodea/com/udabake/model/RecipeIngredientModel.java
@@ -8,7 +8,7 @@ import javax.annotation.Generated;
 public class RecipeIngredientModel {
 
 	@SerializedName("quantity")
-	private int quantity;
+	private float quantity;
 
 	@SerializedName("measure")
 	private String measure;
@@ -16,11 +16,11 @@ public class RecipeIngredientModel {
 	@SerializedName("ingredient")
 	private String ingredient;
 
-	public void setQuantity(int quantity){
+	public void setQuantity(float quantity){
 		this.quantity = quantity;
 	}
 
-	public int getQuantity(){
+	public float getQuantity(){
 		return quantity;
 	}
